+++ .gitignore
@@ -1,4 +1,5 @@
 # Binaries for programs and plugins
+*.idea
 *.yaml.swp
 *.exe
 *.exe~
