--- label/label.go
@@ -1,98 +1,0 @@
-package label
-
-import (
-	"bytes"
-	"encoding/json"
-	"fmt"
-	"io/ioutil"
-	"log"
-	"net/http"
-	"os"
-	"strings"
-	"time"
-)
-
-var (
-	ghRepoOwner                  = os.Getenv("GITHUB_REPOSITORY_OWNER")
-	ghRef                        = os.Getenv("GITHUB_REF")
-	ghRefName                    = os.Getenv("GITHUB_REF_NAME")
-	ghRepo                       = os.Getenv("GITHUB_REPOSITORY")
-	ghToken                      = os.Getenv("GITHUB_TOKEN")
-	ghEvent                      = os.Getenv("GITHUB_EVENT_NAME")
-	ghActor                      = os.Getenv("GITHUB_ACTOR")
-	ghWorkflow                   = os.Getenv("GITHUB_WORKFLOW")
-	ghActionsIDTokenRequestURL   = os.Getenv("ACTIONS_ID_TOKEN_REQUEST_URL")
-	ghActionsIDTokenRequestToken = os.Getenv("ACTIONS_ID_TOKEN_REQUEST_TOKEN")
-	ghAPIURL                     = os.Getenv("GITHUB_API_URL")
-)
-
-func checkEnv() {
-	if ghEvent != "pull_request" {
-		fmt.Println("Error: Not a pull request")
-		os.Exit(2)
-	}
-	if ghRefName == "" {
-		fmt.Println("Error: Github Reference Name not available")
-		os.Exit(2)
-	}
-}
-
-func postLabel(label Label) {
-	requestBody := parseLabel(label)
-	prNumber := strings.Split(ghRefName, "/")[0]
-	url := ghAPIURL + "/repos/" + ghRepo + "/issues/" + prNumber + "/labels"
-	fmt.Println("URL: ", url)
-
-	timeout := time.Duration(5 * time.Second)
-	client := &http.Client{
-		Timeout: timeout,
-	}
-
-	request, err := http.NewRequest("POST", url, bytes.NewBuffer(requestBody))
-	request.Header.Add("Accept", "application/vnd.github.v3+json")
-	request.Header.Add("Authorization", "token "+ghToken)
-	request.Header.Set("Content-Type", "application/json")
-	if err != nil {
-		log.Fatalln(err)
-	}
-
-	resp, err := client.Do(request)
-	if err != nil {
-		log.Fatalln(err)
-	}
-	defer resp.Body.Close()
-
-	body, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		log.Fatalln(err)
-	}
-	log.Printf(string(body))
-	fmt.Println()
-}
-
-type arrayFlags []string
-
-func(i *arrayFlags) String() string {
-	return ""
-
-}
-
-func(i *arrayFlags) Set(value string) error {
-	*i = append(*i, strings.TrimSpace(value))
-	return nil
-}
-// colors
-// orange : #D93F0B
-type Label struct {
-	Name 					[]string 	`json:"labels"`
-	Description 	string 	`json:"description,omitempty"`
-	Color					string  `json:"color,omitempty"`
-}
-func parseLabel(label Label) []byte {
-	rb, err := json.Marshal(label)
-	if err != nil {
-		log.Fatalln(err)
-	}
-	fmt.Println(string(rb))
-	return rb
-}
